replace int by same type (#318) (#327)

+++ examples/cpp/source/optimization_solvers/saga_logistic_loss_dense_batch.cpp
@@ -69,7 +69,7 @@ int main(int argc, char * argv[])
     const size_t nParameters = (nFeatures + 1);
     float argument[nParameters];
 
-    for (int i = 0; i < nParameters; i++) argument[i] = 0.f;
+    for (size_t i = 0; i < nParameters; i++) argument[i] = 0.f;
     argument[0] = 0.f;
     argument[1] = 0.f;
